requirements.txt
django==3.2
django-guardian==2.3.0